requirements.txt
# Core
fastapi
uvicorn[standard]
python-dotenv
python-multipart

# AI / LLM
google-genai

# Search & Crawling
ddgs
crawl4ai
playwright

# RAG / Vector
chromadb
pypdf

# Data
pydantic
